chore(clean): move reusable util function to common
